backend/services/agent-dendritic/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
nats-py==2.6.0
pydantic==2.5.2
python-dotenv==1.0.0
